Fix all current compiler warnings on FreeBSD/x86-64.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36038 72102866-910b-0410-8b05-ffd578937521

+++ ColorPickers/GSNamedColorPicker.m
@@ -175,13 +175,13 @@
 }
 
 -(NSString *) browser: (NSBrowser *)sender 
-	titleOfColumn: (int)column
+	titleOfColumn: (NSInteger)column
 {
   return nil;
 }
 
 -(void) browser: (NSBrowser *)sender 
-createRowsForColumn: (int)column
+createRowsForColumn: (NSInteger)column
        inMatrix: (NSMatrix *)matrix
 {
   int i;
@@ -215,16 +215,16 @@ createRowsForColumn: (int)column
 } 
 
 - (BOOL) browser: (NSBrowser*)sender 
-       selectRow: (int)row
-	inColumn: (int)column
+       selectRow: (NSInteger)row
+	inColumn: (NSInteger)column
 {
   return NO;
 }
 
 - (void) browser: (NSBrowser *)sender 
  willDisplayCell: (id)cell
-	   atRow: (int)row 
-	  column: (int)column
+	   atRow: (NSInteger)row 
+	  column: (NSInteger)column
 {
 }
