+++ Source/NSHelpPanel.m
@@ -130,7 +130,7 @@ static NSHelpPanel	*_sharedPanel = nil;
 }
 
 - (id) initWithContentRect: (NSRect)contentRect
-		 styleMask: (unsigned int)aStyle
+		 styleMask: (NSUInteger)aStyle
 		   backing: (NSBackingStoreType)bufferingType
 		     defer: (BOOL)flag
 {
